Print rooms, exits, prompt, & navigate.
Rooms are a nested map, with the keys being keywords uniquely
identifying each room.

The prompt part is a bit ugly, but it mostly works. Need to eliminate
the extra parens around exits.

Still haven't sorted out how I want to do commands. It shouldn't be in
mainloop, just so we can test it better. Some handle-cmd function would
then return either the new state (new room, that is), some way to
no-op, or some way to quit.